requirements.txt
pymodbus>=3.0.0
paho-mqtt>=1.6.1
PyYAML>=6.0
